Lesson 8. Authentication and validation edit/delete posts

+++ app/Http/Controllers/PostController.php
@@ -10,6 +10,12 @@ use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index','show');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -28,8 +34,9 @@ class PostController extends Controller
         }
 
         $posts = Post::join('users', 'author_id', '=', 'users.id')
-            ->orderBy('posts.created_at', 'desc')
+            ->orderBy('posts.created_at', 'posts.desc')
             ->paginate(4);
+
         return view('posts.index', compact('posts'));
     }
 
@@ -55,7 +62,7 @@ class PostController extends Controller
         $post->title = $request->title;
         $post->short_title = Str::length($request->title) > 30 ? Str::substr($request->title, 0, 30) . "..." : $request->title;
         $post->descr = $request->descr;
-        $post->author_id = rand(1, 4);
+        $post->author_id = \Auth::user()->id;
 
         if ($request->file('img')) {
             $path = Storage::putFile('public', $request->file('img'));
@@ -78,8 +85,13 @@ class PostController extends Controller
      */
     public function show($id)
     {
-        $post = Post::join('users', 'author_id', '=', 'users.id')
-            ->find($id);
+        $post = Post::join('users', 'author_id', '=', 'users.id')->select('posts.*','users.name')
+                ->find($id);
+
+        if(!$post){
+            return redirect()->route('post.index')->withErrors('Path not found!');
+        }
+
         return view('posts.show', compact('post'));
     }
 
@@ -92,6 +104,15 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(!$post){
+            return redirect()->route('post.index')->withErrors('Path not found!');
+        }
+
+        if($post->author_id != \Auth::user()->id){
+            return redirect()->route('post.index')->withErrors('You can not edit this post!');
+        }
+
         return view('posts.edit', compact('post'));
     }
 
@@ -105,6 +126,15 @@ class PostController extends Controller
     public function update(PostRequest $request, $id)
     {
         $post = Post::find($id);
+
+        if(!$post){
+            return redirect()->route('post.index')->withErrors('Path not found!');
+        }
+
+        if($post->author_id != \Auth::user()->id){
+            return redirect()->route('post.index')->withErrors('You can not edit this post!');
+        }
+
         $post->title = $request->title;
         $post->short_title = Str::length($request->title) > 30 ? Str::substr($request->title, 0, 30) . "..." : $request->title;
         $post->descr = $request->descr;
@@ -116,7 +146,7 @@ class PostController extends Controller
         }
 
         $post->update();
-        $id = $post->id;
+        $id = $post->post_id;
         return redirect()->route('post.show',compact('id'))->with('success', 'Post updated successfully!');
     }
 
@@ -129,6 +159,15 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if($post->author_id != \Auth::user()->id){
+            return redirect()->route('post.index')->withErrors('You can not delete this post!');
+        }
+
+        if(!$post){
+            return redirect()->route('post.index')->withErrors('Path not found!');
+        }
+
         $post->delete();
         return redirect()->route('post.index')->with('success', 'Post deleted successfully!');
     }
